agent/requirements.txt
psutil==5.9.5
requests==2.32.3
python-dotenv==1.0.0
